Use unsigned long for game id

+++ SteamOverlay/Settings/Models/InjectorConfig.cs
@@ -5,7 +5,7 @@ namespace SteamOverlay.Settings.Models
         public string SteamDir { get; set; }
         public string ProcessName { get; set; }
         public string WorkingDir { get; set; }
-        public int GameId { get; set; }
+        public ulong GameId { get; set; }
         public bool ENABLE_VK_LAYER_VALVE_steam_overlay_1 { get; set; }
     }
 }
